Resolve lodash-related vulnerabilities and suppress Cypress
Configured audit-ci to allow suppressing selected vulnerabilities.

Since audit-ci allows us to suppress reported vulnerabilities when we
need to (e.g. when it's on a dev dependency and there's no fix yet) it
makes sense to use it in the pipeline and ship script instead of NPM's built in audit
command.